+++ test/platform/falconmetrics_flutter_method_channel_test.dart
@@ -1,5 +1,5 @@
 import 'package:falconmetrics_flutter/src/events.dart';
-import 'package:falconmetrics_flutter/src/event_proto_converter.dart';
+import 'package:falconmetrics_flutter/src/platform/event_proto_converter.dart';
 import 'package:falconmetrics_flutter/src/falconmetrics_flutter_method_channel.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';